refactor: created new file for each new widget exported paths in molecules.dart
